Avoid re-adding brew to system PATH
The setup.sh script now checks if brew is already in the system PATH before attempting to add it. This helps prevent any redundancy and potential errors that might happen due to attempting to add brew to the system PATH multiple times. The change involves checking the '~/.zprofile' file before updating.