Ensure zprofile and zshrc files exist before grep
Changes to the setup.sh script now ensure that the ~/.zprofile and ~/.zshrc files exist before attempting to grep them. This will prevent potential errors when these files are not present. This consideration was added as part of the checks for already existing brew in system PATH, and custom shell setup in zshrc.