Update condition in setup.sh script
The condition has been updated in the setup.sh script for detecting a custom shell setup. Previously, it was checking for 'neofetch' in the ~/.zshrc file. Now it looks for the 'kill_it_with_fire_before_it_lays_eggs' string. Additionally, the call to 'neofetch' was removed from the end of the script.