Move Xcode installation steps to setup.sh
The instructions to install Xcode have been taken out of the README.md file and moved into the setup.sh script. This allows for a more streamlined experience, as code execution is automated during installation, and the user is provided with feedback when the Command Line Tools for Xcode are already installed.